Update bip-0070-payment-protocol.py
Included place holders for SSL certificate chains.
It is worth noting that the 'paymentrequest' function is currently dependant on the 'pycrypto' library which can be installed using pip but it would be more elegant to use 'hashlib' if possible.
Also the 'sds_pr' and 'sds_pa' variables are written to the physical disk which is undesirable. It would make more sense to use 'BytesIO' or 'StringIO' from the 'io' library (which ever is appropriate).