Add support for Core's new script_(in)valid.json format
As of Bitcoin Core v0.12.0 script tests can now specify what script evaluation
flags should be applied, and in addition, a dummy transaction is created so
CHECKSIG and the like can be tested.

The .json test files are copied from Bitcoin v0.12.0, but as we don't yet have
support for script evaluation flags the test cases that fail due to that lack
of support have been removed for now; they can be added back later.