Rename and document hexlify/unhexlify wrappers
These wrapper work differently than functions from binascii,
even though they have the same name. That could be a source
of confusion. In this commit, the functions are renamed to
make it clear that they operate with strings.

Because they cannot handle Unicode in any meaningful way,
I have changed their implementation to not even attempt it,
as it is more complicated to handle Unicode, and it doesn't
even work anyway.